Including new kinfuLS_rgb_invdepth
Fork of PCL, including a modified version of kinfu_large_scale which uses the method proposed in [1] instead of ICP for the camera tracking or visual odometry. It is a direct method which computes the visual odometry by minimising both the photometric and geometric error, based on inverse depth. The rest of the functions used in the original kinfu_large_scale implementation for raycasting, volume integration and volume shifting are kept unchanged.

[1] D. Gutiérrez-Gómez, W. Mayol-Cuevas, J.J. Guerrero. "Inverse Depth for Accurate Photometric and Geometric Error Minimisation in RGB-D Dense Visual Odometry", In International Conference on Robotics and Automation (ICRA), to appear, 2015.